requirements.txt
fastapi
crewai
dotenv-python
uvicorn
sqlalchemy
psycopg2-binary
alembic
rq
